New model: c997e54e-873d-4f90-be14-df5af1063af4/650

+++ selfdrive/modeld/models/nav.h
@@ -65,7 +65,7 @@ constexpr int NAV_NET_OUTPUT_SIZE = NAV_OUTPUT_SIZE - 1;
 
 struct NavModelState {
   RunModel *m;
-  float net_input_buf[NAV_INPUT_SIZE];  // TODO: make this uint8_t
+  uint8_t net_input_buf[NAV_INPUT_SIZE];
   float output[NAV_OUTPUT_SIZE];
 };
